+++ Session.php
@@ -11,6 +11,7 @@ use Yii;
 use yii\base\ErrorHandler;
 use yii\base\InvalidConfigException;
 use yii\di\Instance;
+use yii\web\MultiFieldSession;
 
 /**
  * Session extends [[\yii\web\Session]] by using MongoDB as session data storage.
@@ -30,12 +31,15 @@ use yii\di\Instance;
  * ]
  * ~~~
  *
+ * Session extends [[MultiFieldSession]], thus it allows saving extra fields into the [[sessionCollection]].
+ * Refer to [[MultiFieldSession]] for more details.
+ *
  * @property boolean $useCustomStorage Whether to use custom storage. This property is read-only.
  *
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.0
  */
-class Session extends \yii\web\Session
+class Session extends MultiFieldSession
 {
     /**
      * @var Connection|array|string the MongoDB connection object or the application component ID of the MongoDB connection.
@@ -61,16 +65,6 @@ class Session extends \yii\web\Session
     {
         parent::init();
         $this->db = Instance::ensure($this->db, Connection::className());
-    }
-
-    /**
-     * Returns a value indicating whether to use custom session storage.
-     * This method overrides the parent implementation and always returns true.
-     * @return boolean whether to use custom storage.
-     */
-    public function getUseCustomStorage()
-    {
-        return true;
     }
 
     /**
@@ -102,10 +96,7 @@ class Session extends \yii\web\Session
             }
         } else {
             // shouldn't reach here normally
-            $collection->insert([
-                'id' => $newID,
-                'expire' => time() + $this->getTimeout()
-            ]);
+            $collection->insert($this->composeFields($newID, ''));
         }
     }
 
@@ -118,14 +109,20 @@ class Session extends \yii\web\Session
     public function readSession($id)
     {
         $collection = $this->db->getCollection($this->sessionCollection);
+        $condition = [
+            'id' => $id,
+            'expire' => ['$gt' => time()],
+        ];
+
+        if (isset($this->readCallback)) {
+            $doc = $collection->findOne($condition);
+            return $doc === null ? '' : $this->extractData($doc);
+        }
+
         $doc = $collection->findOne(
-            [
-                'id' => $id,
-                'expire' => ['$gt' => time()],
-            ],
+            $condition,
             ['data' => 1, '_id' => 0]
         );
-
         return isset($doc['data']) ? $doc['data'] : '';
     }
 
@@ -143,11 +140,7 @@ class Session extends \yii\web\Session
         try {
             $this->db->getCollection($this->sessionCollection)->update(
                 ['id' => $id],
-                [
-                    'id' => $id,
-                    'data' => $data,
-                    'expire' => time() + $this->getTimeout(),
-                ],
+                $this->composeFields($id, $data),
                 ['upsert' => true]
             );
         } catch (\Exception $e) {